# specs/001-purple-glow-social/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Purple Glow Social API
  version: 1.0.0
  description: |
    RESTful API for Purple Glow Social - AI-powered social media management platform.

    **Authentication**: All endpoints (except /auth/*) require Bearer token in Authorization header.

    **Rate Limiting**: 
    - Authenticated: 100 req/min per user
    - Unauthenticated: 10 req/min per IP

    **Error Responses**: All endpoints return RFC 7807 Problem Details for HTTP APIs format.
  contact:
    name: Purple Glow Social Support
    email: support@purpleglowsocial.com
  license:
    name: Proprietary
servers:
  - url: https://api.purpleglowsocial.com/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
  schemas:
    Error:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: 'https://api.purpleglowsocial.com/errors/validation-error'
        title:
          type: string
          example: 'Validation Error'
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: 'The request body contains invalid data.'
        instance:
          type: string
          format: uri
          example: '/api/v1/posts'
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    SocialMediaAccount:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_profile_id:
          type: string
          format: uuid
        platform:
          type: string
          enum: [facebook, instagram, twitter, linkedin]
        platform_user_id:
          type: string
        platform_username:
          type: string
        connection_status:
          type: string
          enum: [connected, expired, revoked, error]
        last_sync_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_profile_id:
          type: string
          format: uuid
        caption:
          type: string
        language:
          type: string
        image_url:
          type: string
          format: uri
          nullable: true
        image_prompt:
          type: string
          nullable: true
        hashtags:
          type: array
          items:
            type: string
        platform_targets:
          type: array
          items:
            type: string
            enum: [facebook, instagram, twitter, linkedin]
        status:
          type: string
          enum: [pending, approved, scheduled, publishing, published, failed, rejected]
        scheduled_time:
          type: string
          format: date-time
          nullable: true
        published_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user_edits:
          type: array
          items:
            type: object
        ai_model_version:
          type: string
          nullable: true
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: 'https://api.purpleglowsocial.com/errors/unauthorized'
            title: 'Unauthorized'
            status: 401
            detail: 'Invalid or expired authentication token.'
...
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: 'https://api.purpleglowsocial.com/errors/not-found'
            title: 'Not Found'
            status: 404
            detail: 'The requested resource does not exist.'
...
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: 'https://api.purpleglowsocial.com/errors/rate-limit'
            title: 'Rate Limit Exceeded'
            status: 429
            detail: 'You have exceeded the rate limit. Try again in 60 seconds.'
security:
  - BearerAuth: []
paths:
...
  /social-accounts:
    get:
      tags:
        - Social Accounts
      summary: List connected social accounts
      responses:
        '200':
          description: Social accounts retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/SocialMediaAccount'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /social-accounts/{id}:
    delete:
      tags:
        - Social Accounts
      summary: Disconnect social account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Account disconnected
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ========================================
  # Post Endpoints
  # ========================================
  /posts:
    get:
      tags:
        - Posts
      summary: List posts
      description: Get posts for current business profile with filtering
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, scheduled, published, failed, rejected]
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: platform
          in: query
          schema:
            type: string
            enum: [facebook, instagram, twitter, linkedin]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Posts retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      total_pages:
                        type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Posts
      summary: Generate new post
      description: Use AI to generate post caption and image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - platform_targets
              properties:
                topic:
                  type: string
                  description: Optional topic override (uses profile topics if not provided)
                platform_targets:
                  type: array
                  items:
                    type: string
                    enum: [facebook, instagram, twitter, linkedin]
                  minItems: 1
                generate_image:
                  type: boolean
                  default: true
                image_prompt:
                  type: string
                  description: Optional image generation prompt override
      responses:
        '201':
          description: Post generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Validation error or post limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...